refactor MarketPage
