+++ src/pages/MarketPage/components/BondsTable/TableCells/SizeCell.tsx
